# proxy_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Auth-service gateway.",
        "title": "Auth service API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Majid Darvishan",
            "url": "https://github.com/shpd",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/api",
    "paths": {
        "/admin/custom-endpoints": {
            "post": {
                "description": "Create Custom Endpoint to redirect its requests to another endpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create Custom Endpoint",
                "parameters": [
                    {
                        "description": "CustomEndpoint payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerCustomEndpoint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during token generation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.SwaggerCustomEndpoint": {
            "description": "SwaggerCustomEndpoint defines the expected request body for custom endpoint.",
            "type": "object",
            "properties": {
                "endpoints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "needAccounting": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
    }
}